i am ready to change structure of code
